style: melhora no visual da aplicação e organização dos componentes melhorada
