+++ src/Components/Pokemon/Pokemon.tsx
@@ -1,9 +1,9 @@
 import React from 'react'
 import './Pokemon.css'
-import Delete from './Delete'
+import Delete from '../Delete'
 import { useDispatch } from 'react-redux'
-import { addToPokedex, removeFromPokedex } from '../redux/pokedex/actions'
-import Star from './Star';
+import { addToPokedex, removeFromPokedex } from '../../redux/pokedex/actions'
+import Star from '../Star'
 
 interface PokemonProps {
 	name: string
@@ -19,7 +19,7 @@ const Pokemon = ({name, image_url, types, id, favorite = false}: PokemonProps): 
 	return <div className="Pokemon">
 		{!favorite? <Delete onClick={() => {dispatch(removeFromPokedex(id))}} className="close" /> :
 			<Star onClick={() => {dispatch(addToPokedex({name, image_url, types, id}))}} className="close" />}
-		<img src={image_url} alt=""/>
+		<img src={image_url} alt={`${name}`} />
 		<div className="info">
 			<div className="name">{name}<span>#{id}</span></div>
 			<div className="species">{species}</div>
