+++ src/Components/Pokemon/index.ts
@@ -1,0 +1,3 @@
+import Pokemon from './Pokemon'
+
+export default Pokemon
