+++ src/Components/Search.tsx
@@ -1,0 +1,13 @@
+import React from 'react'
+
+interface SearchProps {
+	onClick(): void
+}
+
+const Search = ({onClick}: SearchProps) => (
+	<svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke="currentColor" onClick={onClick}>
+		<path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M21 21l-6-6m2-5a7 7 0 11-14 0 7 7 0 0114 0z" />
+	</svg>
+)
+
+export default Search
