--- src/Containers/Pokedex.css
@@ -1,4 +1,0 @@
-.Pokedex {
-  display: flex;
-  flex-wrap: wrap;
-}
