+++ src/Containers/Pokedex/index.ts
@@ -1,0 +1,2 @@
+import Pokedex from './Pokedex'
+export default Pokedex
