+++ src/Containers/SearchForm/SearchForm.css
@@ -1,0 +1,30 @@
+.App-header {
+  display: flex;
+  flex-direction: column;
+}
+
+.App-header .title {
+  font-family: 'Press Start 2P', cursive;
+  font-size: 40px;
+  padding: 30px;
+  font-weight: 400;
+}
+
+.App-header input {
+  border: 1px solid black;
+  border-radius: 4px;
+  padding: 5px;
+  text-align: center;
+}
+
+.SearchForm {
+  display: flex;
+}
+.SearchForm input{
+  display: flex;
+  flex-grow: 1;
+}
+.SearchForm svg {
+  margin-left: 5px;
+  width: 20px;
+}
