+++ src/Containers/SearchForm/index.ts
@@ -1,0 +1,2 @@
+import SearchForm from './SearchForm'
+export default SearchForm
