+++ src/Containers/SearchResults/SearchResults.css
@@ -1,4 +1,5 @@
 .SearchResult {
   display: flex;
   flex-wrap: wrap;
+  justify-content: center;
 }
