+++ src/redux/reducer.ts
@@ -1,6 +1,6 @@
 import { combineReducers } from 'redux'
-import { pokedexReducer } from './pokedex/reducer';
-import searchReducer from './search/reducer';
+import { pokedexReducer } from './pokedex/reducer'
+import searchReducer from './search/reducer'
 
 const rootReducer = combineReducers({pokedexReducer, searchReducer})
